requirements.txt
fastapi
uvicorn[standard]
httpx
requests
python-dotenv
pydantic
pymongo
cohere
tqdm
PyJWT
google-cloud-translate==3.13.0
groq
google-auth==2.28.1
aiohttp
asyncio
langdetect
markdown
# openai
# langchain
# chromadb
# pgvector
# sqlalchemy
# psycopg2-binary
# googletrans==4.0.0rc1
# transformers
# sentence-transformers
# nltk
# passlib[bcrypt]
# python-multipart
# email-validator
# python-jose[cryptography]
# indic-trans
# celery
# redis